requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
geoip2==4.7.0
idna==3.4
maxminddb==2.5.1
multidict==6.0.4
oyaml==1.0
pystun3==1.0.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
timeparse-plus==1.2.0
types-python-dateutil==2.8.19.14
urllib3==2.1.0
websockets==12.0
yarl==1.9.2
